requirements.txt
grpcio==1.49.1
grpcio-tools==1.49.1
loguru==0.6.0